requirements.txt
python==3.11.8
pytorch-gpu==2.1.2
torchvision==0.16.2
timm==0.9.16
numpy==1.24.3
scikit-learn==1.3.0
pandas==2.1.1
matplotlib==3.8.0
kornia==0.7.1
